% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rubric_item.R
\name{add_rubric_item}
\alias{add_rubric_item}
\title{Prompts grader to add a rubric item for the current question}
\usage{
add_rubric_item(curr_q, rubric_path, rubric_list)
}
\arguments{
\item{curr_q}{string, the name of the question that was being graded and for which the grader has asked to add a rubric item. Needs to be one of the names in the column}

\item{rubric_path}{string, path to assignment rubric. This rubric should be created using the function create_rubric_template, then filled in by the user. The rubric file name and column names must not be changed.}

\item{rubric_list}{list whose format corresponds to rubric_list, which is used by most functions in this package. This is produced by import_rubric}
}
\value{
rubric list
}
\description{
Prompts grader to add a rubric item for the current question
}
