% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_grade_write_feedback.R
\name{assign_grade_write_feedback}
\alias{assign_grade_write_feedback}
\title{Writes feedback, and returns grade, grade decomposition, and grading status}
\usage{
assign_grade_write_feedback(temp_grade_sheet_row, rubric_list, rubric_prompts)
}
\arguments{
\item{temp_grade_sheet_row}{data frame with 1 row; The temp_grade_sheet is a data frame containing information for gradetools internal use. This row should correspond to the student that needs to be graded and have feedback written}

\item{rubric_list}{list of lists; each sub-list corresponds to a part of the assignment rubric and specifies the name of this part (e.g. "question_1"), the total points that can be earned for this part, the rubric etc. Rubric should be made using 'create_rubric_template()'.}

\item{rubric_prompts}{list of prompts; One prompt for each question plus one for overall feedback. This is produced by create_rubric_prompts}
}
\value{
list with\describe{
  \item{grade}{A number, the total grade scored for this assignment by the student being graded (e.g. 3)}
  \item{grade_decomposition}{A vector that shows how each part of the assignment contributed to the overall grade (e.g. c(1, 0, 2))}
  \item{grading_status}{A string indicating the grading_status for this student}
}
}
\description{
Writes feedback, and returns grade, grade decomposition, and grading status
}
\keyword{internal}
