% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist_regrading.R
\name{assist_regrading}
\alias{assist_regrading}
\title{Assists with regrading, given assist_grading() or assist_advanced_grading() was previously used}
\usage{
assist_regrading(
  rubric_path,
  temp_grade_sheet_path,
  final_grade_sheet_path,
  questions_to_regrade,
  students_to_regrade,
  teams_to_regrade = NULL,
  missing_assignment_grade = NA,
  github_issues = FALSE
)
}
\arguments{
\item{rubric_path}{string, path to assignment rubric. 
This rubric should be created using the function create_rubric_template, then filled in by the user. 
The rubric file name and column names must not be changed.}

\item{temp_grade_sheet_path}{string; assist-grading() functions save a file which includes information for gradetools's internal use. 
This is that path for that file. Must be a .csv}

\item{final_grade_sheet_path}{string; path to save final grade sheet to. Must be a .csv}

\item{questions_to_regrade}{vector of strings; vector of assignment question 'names' to regrade. This can be specified as "all" to apply to all questions. questions_to_grade must exactly match ones present in the rubric}

\item{students_to_regrade}{vector of strings; vector of student identifiers to regrade. This can be specified as "all" to apply to all student's in the rubric. students_to_grade must be student_identifiers present in the roster. This should not be provided if the assignment involved team grading}

\item{teams_to_regrade}{vector of strings; This argument is for team grading (i.e. if one assignment is connected with multiple students). team_identifiers corresponding to teams to grade, or "all" to specify all assignments should be graded. All teams_to_regrade must be team_identifiers present in the roster}

\item{missing_assignment_grade}{numeric; The grade to assign a student with no assignment submission}

\item{github_issues}{logical, whether the grader wants to be given the option to create an issue in students' repos or not (defaults to FALSE)}
}
\description{
Assists with regrading, given assist_grading() or assist_advanced_grading() was previously used
}
