% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_final_grade_sheet.R
\name{create_final_grade_sheet}
\alias{create_final_grade_sheet}
\title{Writes final grade sheet csv}
\usage{
create_final_grade_sheet(
  grading_progress_log,
  final_grade_sheet_path,
  missing_assignment_grade,
  rubric_list,
  rubric_prompts,
  team_grading,
  write_grades_into_feedback
)
}
\arguments{
\item{grading_progress_log}{data frame; The grading_progress_log is a data frame containing information for gradetools internal use}

\item{final_grade_sheet_path}{string; path to save final grade sheet to. Must be a .csv}

\item{missing_assignment_grade}{numeric; The grade to assign a student with no assignment submission}

\item{rubric_list}{list whose format corresponds to rubric_list, which is used by most functions in this package. This is produced by import_rubric}

\item{rubric_prompts}{list of prompts; One prompt for each question plus one for overall feedback. This is produced by create_rubric_prompts}

\item{team_grading}{logical, indicates if any assignment submission is associated with multiple students (e.g. team projects)}

\item{write_grades_into_feedback}{logical, whether to write numeric grades into the feedback file, along with qualitative feedback (defaults to FALSE).}
}
\description{
Writes final grade sheet csv
}
\keyword{internal}
