% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_to_github.R
\name{create_issues_github}
\alias{create_issues_github}
\title{Push issues to GitHub}
\usage{
create_issues_github(temp_grade_sheet, class_github_name, team_grading = FALSE)
}
\arguments{
\item{temp_grade_sheet}{tibble created by assist_advanced_grading().}

\item{class_github_name}{string, GitHub name of the class. This input is needed to push feedback and issues to the correct class on GitHub.}

\item{team_grading}{logical, indicates if any assignment submission is associated with multiple students (e.g. team projects)}
}
\description{
Push issues to GitHub
}
