% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rubric_template.R
\name{create_rubric_template}
\alias{create_rubric_template}
\title{Creates rubric template csv file}
\usage{
create_rubric_template(
  rubric_path,
  negative_grading = TRUE,
  create_example_rubric = FALSE,
  example_rubric_path = NULL
)
}
\arguments{
\item{rubric_path}{string; path used to save the rubric template. This must be a csv file.}

\item{negative_grading}{Boolean; TRUE indicates the assignment will use negative grading instead of positive.}

\item{create_example_rubric}{Boolean; TRUE indicates an example rubric will also be created.}

\item{example_rubric_path}{string; path used to save the example rubric if create_example_rubric is TRUE. This must be a csv file.}
}
\description{
Creates rubric template csv file
}
\details{
The rubric must be specified in the following way: 
    The column names generated must be left the same. 
    Each row of the rubric will contain the details for one rubric item for one question. 
    Every question to be graded must have at least one rubric item associated.
    The entries in the 'name' column must be 'general_feedback', 'all_questions', or the name of the question that the rubric item is specified for.
    Rubric items with 'name' specified as 'all_questions' are rubric items available for all assignment questions. There should be no total_points value for these rubric items. The points_to_remove or points_to_add values will be taken as percentages (e.g. If points_to_remove for an all_questions rubric item is specified as 25, when the grader chooses this grade rubric option while grading 25% of the total_points for that question will be deducted.).
    Rubric items with 'name' specified as 'general_feedback' are rubric items which the user will be prompted for after they have completely graded an assignment. These rubric items do not have any points attached, their purpose is to leave a general message to the student after grading has been completed.
    All rubric prompt codes must begin with a digit, contain no spaces, and be unique within a question. 
    Rubric prompt messages are brief summaries to remind the user of the appropriate situation to apply that rubric item.
    When the user is grading they will see see all rubric items for that question and 'all_questions' rubric items. The corresponding prompt codes points to remove/add and prompt message will be displayed.
    The 'feedback' column contains the message that will be printed in the student's feedback file if that rubric item is selected.
}
