% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_feedback_grade.R
\name{delete_feedback_grade}
\alias{delete_feedback_grade}
\title{Delete feedback and grades for specified students and questions}
\usage{
delete_feedback_grade(
  temp_grade_sheet_path,
  rubric_path,
  student_ids,
  questions_to_delete
)
}
\arguments{
\item{temp_grade_sheet_path}{string; path to temporary grade sheet produced by assign_grade}

\item{rubric_path}{string, path to assignment rubric. This rubric should be created using the function create_rubric_template, then filled in by the user. The rubric file name and column names must not be changed.}

\item{student_ids}{string; vector of student identifiers to have grade and feedback deleted. This can be specified as "all" to apply to all student's in the rubric}

\item{questions_to_delete}{string; vector of assignment question 'names' to have grade and feedback deleted. This can be specified as "all" to apply to all questions.}
}
\description{
Delete feedback and grades for specified students and questions
}
