% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_student_grading_progress.R
\name{delete_student_grading_progress}
\alias{delete_student_grading_progress}
\title{Delete feedback and grades for specified students and questions}
\usage{
delete_student_grading_progress(
  grading_progress_log_path,
  rubric_path,
  identifier,
  questions_to_delete,
  github_issues
)
}
\arguments{
\item{grading_progress_log_path}{string; path to grading progress log produced by assign_grade}

\item{rubric_path}{string; path to assignment rubric.}

\item{identifier}{string; single student_identifier (or team_identifier) to have grading progress deleted.}

\item{questions_to_delete}{string; vector of assignment question 'names' to have grade and feedback deleted. 
This can be specified as "all" to apply to all questions.}

\item{github_issues}{logical, whether the grader wants to be given the option to create an issue in students' repos or not (defaults to FALSE)}
}
\description{
Delete feedback and grades for specified students and questions
}
\keyword{internal}
