% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_github_issues.R
\name{note_github_issues}
\alias{note_github_issues}
\title{Collects GitHub issue title and body}
\usage{
note_github_issues(curr_row, curr_q)
}
\arguments{
\item{curr_row}{a row extracted from the grading_progress_log data table within assist_grading().}

\item{curr_q}{name of current question}
}
\value{
corresponding row of grading progress log
}
\description{
Collects GitHub issue title and body
}
\keyword{internal}
