% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_student_grading_progress.R
\name{remove_associated_info}
\alias{remove_associated_info}
\title{Helper function for deleting info}
\usage{
remove_associated_info(curr_row, questions_to_delete, question_col, info_cols)
}
\arguments{
\item{curr_row}{string; row of temporary grade sheet to change}

\item{questions_to_delete}{string; vector of assignment question 'names' to have grade and feedback deleted.}

\item{question_col}{string; name of column containing question names for associated info}

\item{info_cols}{vector of strings; names of columns containing info to be deleted}
}
\description{
Helper function for deleting info
}
\keyword{internal}
