% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_feedback.R
\name{render_feedback}
\alias{render_feedback}
\title{Render feedback files for graded students}
\usage{
render_feedback(grading_progress_log)
}
\arguments{
\item{grading_progress_log}{The grading progress log used internally by gradetools}
}
\description{
Render feedback files for graded students
}
\keyword{internal}
