% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_github_issues.R
\name{request_github_issues}
\alias{request_github_issues}
\title{Allows the user to input their requested GitHub issues titles and bodies}
\usage{
request_github_issues(curr_row)
}
\arguments{
\item{curr_row}{a row extracted from the temp_grade_sheet data table within assist_grading().}
}
\description{
Allows the user to input their requested GitHub issues titles and bodies
}
