% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_final_grade_sheet.R
\name{create_final_grade_sheet}
\alias{create_final_grade_sheet}
\title{Writes final grade sheet csv}
\usage{
create_final_grade_sheet(
  temp_grade_sheet,
  final_grade_sheet_path,
  missing_assignment_grade,
  rubric_list,
  rubric_prompts,
  team_grading
)
}
\arguments{
\item{temp_grade_sheet}{data frame; The temp_grade_sheet is a data frame containing information for gradetools internal use}

\item{final_grade_sheet_path}{string; path to save final grade sheet to. Must be a .csv}

\item{missing_assignment_grade}{numeric; The grade to assign a student with no assignment submission}

\item{rubric_list}{list whose format corresponds to rubric_list, which is used by most functions in this package. This is produced by import_rubric}

\item{rubric_prompts}{list of prompts; One prompt for each question plus one for overall feedback. This is produced by create_rubric_prompts}

\item{team_grading}{logical, indicates if any assignment submission is associated with multiple students (e.g. team projects)}
}
\description{
Writes final grade sheet csv
}
\keyword{internal}
