% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_to_github.R
\name{push_feedback_github}
\alias{push_feedback_github}
\title{Push feedback to github}
\usage{
push_feedback_github(temp_grade_sheet, class_github_name)
}
\arguments{
\item{temp_grade_sheet}{tibble created by assist_advanced_grading().}

\item{class_github_name}{string, GitHub name of the class. This input is needed to push feedback and issues to the correct class on GitHub.}
}
\description{
Push feedback to github
}
