% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_to_github.R
\name{push_to_github}
\alias{push_to_github}
\title{Push feedback and issues to GitHub}
\usage{
push_to_github(
  grading_progress_log_path,
  class_github_name,
  example_github_repo,
  example_identifier,
  push_feedback = FALSE,
  create_issues = FALSE,
  team_grading = FALSE
)
}
\arguments{
\item{grading_progress_log_path}{string; assist-grading() functions save a file which includes information for gradetools's internal use. 
This is that path for that file. Must be a .csv}

\item{class_github_name}{string, GitHub name of the class. This input is only needed if \code{github_issues} is set to TRUE.}

\item{example_github_repo}{string; example of the name of a student's GitHub repo for this assignment (or team's repo, if \code{team_grading} = TRUE). The \code{example_identifier} must feature in the \code{example_github_repo} and must be the same as the roster identifier for the student (or for the team if \code{team_grading} = TRUE).}

\item{example_identifier}{string; the GitHub username of the student or the name of the team, if \code{team_grading} = TRUE. This needs to be present somewhere in \code{example_github_repo}.}

\item{push_feedback}{logical, whether to push feedback files}

\item{create_issues}{logical, whether to create issues}

\item{team_grading}{logical, indicates if any assignment submission is associated with multiple students (e.g. team projects)}
}
\description{
Push feedback and issues to GitHub
}
